Transition Labels are now Strings.
This creates one problem: We can no longer share structure between
the relational and the state-based representation. (A transition in
the state-based model needs an instance of class Action as label).

The interface for creating state-based models also needs to be updated.
This is why there are errors in the build (which can be ignored for core
functionality).