Prepare exported Promela for enhancing events.
These enhanced events will have a global variable tracking the most recent event.
The variables are visible from a never statement.

The Promela for the send, recv and local events have been modified to be Promela
inline statements. This facilitate additions to the events such as a variable
change before or after the event has taken place. This will allows us to track
these events in never claims.