Added basic test cases for the Promela conversion.
Currently, Spin runs under the test-output folder as it generates extra files as
it runs. This probably needs to be dealt with.

The basic test cases for the invariants involve simple systems.

A single test case runs indefinitely. This is
runSpinSimpleConcurrencyString2Success under CSightMainTests. It currently has a
timeout built in so it is possible to test all the cases.

The GFSM is also output with the CFSM for better debugging.