Fix for a memory leak involving killing the Spin process and timeouts.
Removed -DMEMLIM. This compiler option actually causes the model checker to abort
when it reaches the limit. This is not what we want from the verifier.

Instead, we use -DBFS_DISK to cache to disk if memory usage becomes too high.

The memory leak fix forced us to move our syntax check from getVerifyResult() to verify().
The corresponding test has also been changed in response.

Moved the defined ENDSTATECHECK to earlier in Promela source to keep it with
the rest of the defines. This makes our Java code easier to read.

Added more comments to the Promela generation.

Added a method to clean up files that we generate.